Signal early that communication ended
slt sometimes accumulates tons of connections in CLOSE_WAIT state,
eventually leading to a DoS because no new connections are processed.
Since joinConnection deals with half-connections, we can't just
close the connection since that breaks the other half (see
commit f77d60ce061fb2625a51058b2ee5d053a8d326a2 which handled that)
but for TCP specifically there's API support to close just the writing
side, so do that.
This should help the other side determine that it needn't keep its
own side of the connection open for longer.

Requires a type assertion, but that's better than rewriting all of
slt and go-vhost to make it work with net.TCPConn instances instead
of the generic net.Conn type.